src/DGX1/src/EncoderFineTuning/requirements-Training.txt
transformers==4.44.2
datasets==3.0.0
accelerate==0.34.2
evaluate==0.4.3
bitsandbytes==0.43.3
huggingface-hub==0.24.7
peft==0.12.0
trl==0.10.1
python-dotenv==1.0.1
ruff==0.6.4
mypy==1.11.2
torchinfo==1.8.0
mlflow~=2.17.0rc0
python-dotenv~=1.0.1